Change layout and implement gpt-3.5-turbo
Great news! I was able to successfully implement the gpt-3.5-turbo
model, but not without encountering some errors initially. However,
after doing some investigation, I discovered that the issue was due to
using an outdated version of the open ai model. Once I updated to the
current model version, everything started to work smoothly. Of course,
there is still some work to be done, such as tweaking the model's
parameters and ensuring that the prompts are optimized, but so far the
results are looking promising.